+++ src/bitmap/bitmap_data_writer.cpp
@@ -8,10 +8,68 @@
 #include <squish.h>
 
 namespace Invader {
-    void write_bitmap_data(const GeneratedBitmapData &scanned_color_plate, std::vector<std::byte> &bitmap_data_pixels, std::vector<Parser::BitmapData> &bitmap_data, BitmapUsage usage, BitmapFormat format, BitmapType bitmap_type, bool palettize, bool dither_alpha, bool dither_red, bool dither_green, bool dither_blue) {
+    void write_bitmap_data(const GeneratedBitmapData &scanned_color_plate, std::vector<std::byte> &bitmap_data_pixels, std::vector<Parser::BitmapData> &bitmap_data, BitmapUsage usage, std::optional<BitmapFormat> &format, BitmapType bitmap_type, bool palettize, bool dither_alpha, bool dither_red, bool dither_green, bool dither_blue) {
         using namespace Invader::HEK;
 
         auto bitmap_count = scanned_color_plate.bitmaps.size();
+        
+        // If format is nullopt, automatically determine a format
+        bool automatically_determined_format = !format.has_value();
+        if(automatically_determined_format) {
+            // Height maps and vector maps must always be 32-bit
+            if(usage == BitmapUsage::BITMAP_USAGE_HEIGHT_MAP || usage == BitmapUsage::BITMAP_USAGE_VECTOR_MAP) {
+                format = BitmapFormat::BITMAP_FORMAT_32_BIT;
+            }
+            else {
+                // Determine if we can make it smaller
+                bool is_monochrome = true;
+                bool is_16_bit = true;
+                for(auto &b : scanned_color_plate.bitmaps) {
+                    bool is_a0r5g6b5 = true;
+                    bool is_a1r5g5b5 = true;
+                    bool is_a4r4g4b4 = true;
+                    for(auto &pixel : b.pixels) {
+                        is_monochrome = is_monochrome && (pixel.red == pixel.green && pixel.green == pixel.blue);
+                        is_a0r5g6b5 = is_a0r5g6b5 && (pixel == Pixel::convert_from_16_bit<0,5,6,5>(pixel.convert_to_16_bit<0,5,6,5>()));
+                        is_a1r5g5b5 = is_a1r5g5b5 && (pixel == Pixel::convert_from_16_bit<1,5,5,5>(pixel.convert_to_16_bit<1,5,5,5>()));
+                        is_a4r4g4b4 = is_a4r4g4b4 && (pixel == Pixel::convert_from_16_bit<4,4,4,4>(pixel.convert_to_16_bit<4,4,4,4>()));
+                    }
+                    is_16_bit = is_16_bit && (is_a0r5g6b5 || is_a1r5g5b5 || is_a4r4g4b4);
+                    
+                    // If we don't fit into the 16-bit or monochrome color space, bail!
+                    if(!is_16_bit && !is_monochrome) {
+                        break;
+                    }
+                }
+                
+                if(is_monochrome) {
+                    format = BitmapFormat::BITMAP_FORMAT_MONOCHROME;
+                }
+                else if(is_16_bit) {
+                    format = BitmapFormat::BITMAP_FORMAT_16_BIT;
+                }
+                else {
+                    format = BitmapFormat::BITMAP_FORMAT_32_BIT;
+                }
+            }
+            
+            switch(*format) {
+                case BitmapFormat::BITMAP_FORMAT_32_BIT:
+                    oprintf("Automatically determined format as 32-bit\n");
+                    break;
+                case BitmapFormat::BITMAP_FORMAT_16_BIT:
+                    oprintf("Automatically determined format as 16-bit\n");
+                    break;
+                case BitmapFormat::BITMAP_FORMAT_MONOCHROME:
+                    oprintf("Automatically determined format as monochrome\n");
+                    break;
+                default:
+                    std::terminate();
+            }
+        }
+        
+        oprintf("Found %zu bitmap%s:\n", bitmap_count, bitmap_count == 1 ? "" : "s");
+        
         for(std::size_t i = 0; i < bitmap_count; i++) {
             // Write all of the fields here
             auto &bitmap = bitmap_data.emplace_back();
@@ -39,7 +97,7 @@ namespace Invader {
             // Get the data
             std::vector<std::byte> current_bitmap_pixels(reinterpret_cast<const std::byte *>(bitmap_color_plate.pixels.data()), reinterpret_cast<const std::byte *>(bitmap_color_plate.pixels.data() + bitmap_color_plate.pixels.size()));
             auto *first_pixel = reinterpret_cast<Pixel *>(current_bitmap_pixels.data());
-            bitmap.format = BitmapEncode::most_efficient_format(current_bitmap_pixels.data(), bitmap.width, bitmap.height, bitmap.depth, format, bitmap.type);
+            bitmap.format = BitmapEncode::most_efficient_format(current_bitmap_pixels.data(), bitmap.width, bitmap.height, bitmap.depth, *format, bitmap.type);
 
             // Set the format
             bool compressed = (format == BitmapFormat::BITMAP_FORMAT_DXT1 || format == BitmapFormat::BITMAP_FORMAT_DXT3 || format == BitmapFormat::BITMAP_FORMAT_DXT5);